requirements.txt
langchain>=0.0.267
langchain-community>=0.0.13
langchainhub>=0.1.13
langchain-ollama>=0.0.1
pydantic>=2.0.3
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pypdf
docx2txt
python-dotenv
Flask>=2.0
openai-whisper
torch
edge-tts 